controller_bridge/requirements.txt
numpy==2.2.6
numpy-quaternion==2024.0.12
pyzmq==27.1.0
scipy==1.15.3
